requirements.txt
flask
twilio
python-dotenv
openai
